Splitting named requests into partial class and adding constant for syntax paths
