+++ src/Microsoft.DotNet.Interactive.Http/HttpKernel.cs
@@ -133,9 +133,6 @@ public class HttpKernel :
     {
         var parseResult = HttpRequestParser.Parse(command.Code);
         var requestNodes = parseResult.SyntaxTree.RootNode.ChildNodes.OfType<HttpRequestNode>();
-        //var lastSpan = requestNodes.Select(n => n.Span).LastOrDefault();
-
-
 
         if (command.Parameters.TryGetValue("Document", out var doc))
         {
@@ -143,7 +140,7 @@ public class HttpKernel :
             var lastSpan = parsedDoc.SyntaxTree.RootNode.ChildNodes
                 .OfType<HttpRequestNode>()
                 .FirstOrDefault(n => n.Text == requestNodes.Last().Text)?.Span;
-            if(lastSpan != null)
+            if (lastSpan != null)
             {
                 var docVariableNodes = parsedDoc.SyntaxTree.RootNode.ChildNodes.OfType<HttpVariableDeclarationAndAssignmentNode>();
                 var docVariableNames = docVariableNodes.Where(n => n.Span.Start < lastSpan?.Start).Select(n => n.DeclarationNode?.VariableName).ToHashSet();
@@ -155,7 +152,7 @@ public class HttpKernel :
                     }
                 }
             }
-              
+
         }
         else
         {
