+++ src/Microsoft.DotNet.Interactive.Http/Model/HttpNamedRequest.cs
@@ -38,27 +38,27 @@ internal class HttpNamedRequest
             return node.CreateBindingFailure(HttpDiagnostics.InvalidNamedRequestPath(node.Text));
         }
 
-        if (path[0] != Name)
+        if (path[SyntaxDepth.RequestName] != Name)
         {
             return node.CreateBindingFailure(HttpDiagnostics.InvalidNamedRequestPath(node.Text));
         }
 
 
-        if (path[1] == "request")
+        if (path[SyntaxDepth.RequestOrResponse] == "request")
         {
-            if (path[2] == "body")
+            if (path[SyntaxDepth.BodyOrHeaders] == "body")
             {
                 if (RequestNode.BodyNode is null)
                 {
-                    return node.CreateBindingFailure(HttpDiagnostics.InvalidBodyInNamedRequest(path[0]));
+                    return node.CreateBindingFailure(HttpDiagnostics.InvalidBodyInNamedRequest(path[SyntaxDepth.RequestName]));
                 }
                 else
                 {
-                    if (path[3] == "*")
+                    if (path[SyntaxDepth.RawBodyContent] == "*")
                     {
                         return node.CreateBindingSuccess(RequestNode.BodyNode.Text);
                     }
-                    else if (path[3] == "$")
+                    else if (path[SyntaxDepth.JsonRoot] == "$")
                     {
                         try
                         {
@@ -84,14 +84,14 @@ internal class HttpNamedRequest
                         }
 
                     }
-                    else if (path[3].StartsWith("//"))
+                    else if (path[SyntaxDepth.XmlRoot].StartsWith("//"))
                     {
                         try
                         {
                             var xmlDoc = new XmlDocument();
                             xmlDoc.LoadXml(RequestNode.BodyNode.Text);
 
-                            var xmlNodes = xmlDoc.SelectNodes(path[3].Substring(1));
+                            var xmlNodes = xmlDoc.SelectNodes(path[SyntaxDepth.XmlRoot].Substring(1));
 
                             if (xmlNodes is { Count: 1 })
                             {
@@ -99,7 +99,7 @@ internal class HttpNamedRequest
                             }
                             else
                             {
-                                return node.CreateBindingFailure(HttpDiagnostics.InvalidXmlNodeInNamedRequest(path[3]));
+                                return node.CreateBindingFailure(HttpDiagnostics.InvalidXmlNodeInNamedRequest(path[SyntaxDepth.XmlRoot]));
                             }
 
                         }
@@ -111,26 +111,26 @@ internal class HttpNamedRequest
                     }
                 }
             }
-            else if (path[2] == "headers")
+            else if (path[SyntaxDepth.BodyOrHeaders] == "headers")
             {
                 if (RequestNode.HeadersNode is null)
                 {
-                    return node.CreateBindingFailure(HttpDiagnostics.InvalidHeadersInNamedRequest(path[0]));
+                    return node.CreateBindingFailure(HttpDiagnostics.InvalidHeadersInNamedRequest(path[SyntaxDepth.RequestName]));
                 }
 
-                var headerNode = RequestNode.HeadersNode.HeaderNodes.FirstOrDefault(hn => hn.NameNode?.Text == path[3]);
+                var headerNode = RequestNode.HeadersNode.HeaderNodes.FirstOrDefault(hn => hn.NameNode?.Text == path[SyntaxDepth.HeaderName]);
 
                 if (headerNode is null || headerNode.ValueNode is null)
                 {
-                    return node.CreateBindingFailure(HttpDiagnostics.InvalidHeaderNameInNamedRequest(path[3]));
+                    return node.CreateBindingFailure(HttpDiagnostics.InvalidHeaderNameInNamedRequest(path[SyntaxDepth.HeaderName]));
                 }
 
                 return node.CreateBindingSuccess(headerNode.ValueNode.Text);
             }
         }
-        else if (path[1] == "response")
+        else if (path[SyntaxDepth.RequestOrResponse] == "response")
         {
-            if (path[2] == "body")
+            if (path[SyntaxDepth.BodyOrHeaders] == "body")
             {
                 if (Response.Content is null)
                 {
@@ -138,11 +138,11 @@ internal class HttpNamedRequest
                 }
                 else
                 {
-                    if (path[3] == "*")
+                    if (path[SyntaxDepth.RawBodyContent] == "*")
                     {
                         return node.CreateBindingSuccess(Response.Content.Raw);
                     }
-                    else if (path[3] == "$")
+                    else if (path[SyntaxDepth.JsonRoot] == "$")
                     {
 
                         if (Response.Content.ContentType == null || !(Response.Content.ContentType.StartsWith("application/json")))
@@ -177,7 +177,7 @@ internal class HttpNamedRequest
 
 
                     }
-                    else if (path[3].StartsWith("//"))
+                    else if (path[SyntaxDepth.XmlRoot].StartsWith("//"))
                     {
                         if (Response.Content is null)
                         {
@@ -195,7 +195,7 @@ internal class HttpNamedRequest
                             xmlDoc.LoadXml(Response.Content.Raw);
 
                             //Remove the leading slash
-                            var xmlNodes = xmlDoc.SelectNodes(path[3].Substring(1));
+                            var xmlNodes = xmlDoc.SelectNodes(path[SyntaxDepth.XmlRoot].Substring(1));
 
                             if (xmlNodes is { Count: 1 })
                             {
@@ -203,7 +203,7 @@ internal class HttpNamedRequest
                             }
                             else
                             {
-                                return node.CreateBindingFailure(HttpDiagnostics.InvalidXmlNodeInNamedRequest(path[3]));
+                                return node.CreateBindingFailure(HttpDiagnostics.InvalidXmlNodeInNamedRequest(path[SyntaxDepth.XmlRoot]));
                             }
                         }
                         catch (XmlException)
@@ -214,10 +214,10 @@ internal class HttpNamedRequest
                 }
 
             }
-            else if (path[2] == "headers")
+            else if (path[SyntaxDepth.BodyOrHeaders] == "headers")
             {
 
-                if (Response.Headers.TryGetValue(path[3], out var header) && header is not null)
+                if (Response.Headers.TryGetValue(path[SyntaxDepth.HeaderName], out var header) && header is not null)
                 {
                     //If the path is response.headers.<headerName> and the header value is an array, return the first element
                     if (path.Length == 4)
@@ -235,7 +235,7 @@ internal class HttpNamedRequest
                 }
                 else
                 {
-                    return node.CreateBindingFailure(HttpDiagnostics.InvalidHeaderNameInNamedRequest(path[3]));
+                    return node.CreateBindingFailure(HttpDiagnostics.InvalidHeaderNameInNamedRequest(path[SyntaxDepth.HeaderName]));
                 }
 
             }
@@ -280,4 +280,17 @@ internal class HttpNamedRequest
         }
 
     }
+
+    internal static class SyntaxDepth
+    {
+        //They used to refer to the depth of different elements within the HttpNamedRequest syntax
+        public const int RequestName = 0;
+        public const int RequestOrResponse = 1;
+        public const int BodyOrHeaders = 2;
+        public const int XmlRoot = 3;
+        public const int JsonRoot = 3;
+        public const int HeaderName = 3;
+        public const int RawBodyContent = 3; //Asterisk
+    }
 }
+
